+++ src/resources/v1/boxes/fs.ts
@@ -7,11 +7,14 @@ import { path } from '../../../internal/utils/path';
 
 export class Fs extends APIResource {
   /**
+   * List box files
+   *
    * @example
    * ```ts
-   * const fs = await client.v1.boxes.fs.list('id', {
-   *   path: 'path',
-   * });
+   * const fs = await client.v1.boxes.fs.list(
+   *   'c9bdc193-b54b-4ddb-a035-5ac0c598d32d',
+   *   { path: '/home/user/documents' },
+   * );
    * ```
    */
   list(id: string, query: FListParams, options?: RequestOptions): APIPromise<FListResponse> {
@@ -19,11 +22,14 @@ export class Fs extends APIResource {
   }
 
   /**
+   * Read box file
+   *
    * @example
    * ```ts
-   * const response = await client.v1.boxes.fs.read('id', {
-   *   path: 'path',
-   * });
+   * const response = await client.v1.boxes.fs.read(
+   *   'c9bdc193-b54b-4ddb-a035-5ac0c598d32d',
+   *   { path: '/home/user/documents/config.json' },
+   * );
    * ```
    */
   read(id: string, query: FReadParams, options?: RequestOptions): APIPromise<FReadResponse> {
@@ -31,12 +37,17 @@ export class Fs extends APIResource {
   }
 
   /**
+   * Write box file
+   *
    * @example
    * ```ts
-   * const response = await client.v1.boxes.fs.write('id', {
-   *   content: 'content',
-   *   path: 'path',
-   * });
+   * const response = await client.v1.boxes.fs.write(
+   *   'c9bdc193-b54b-4ddb-a035-5ac0c598d32d',
+   *   {
+   *     content: 'Hello, World!\nThis is file content.',
+   *     path: '/home/user/documents/output.txt',
+   *   },
+   * );
    * ```
    */
   write(id: string, body: FWriteParams, options?: RequestOptions): APIPromise<FWriteResponse> {
@@ -44,14 +55,20 @@ export class Fs extends APIResource {
   }
 }
 
+/**
+ * Response containing directory listing results
+ */
 export interface FListResponse {
   /**
-   * A box instance that can be either Linux or Android type
+   * Array of files and directories
    */
   data: Array<FListResponse.File | FListResponse.Dir>;
 }
 
 export namespace FListResponse {
+  /**
+   * File system file representation
+   */
   export interface File {
     /**
      * Last modified time of the file
@@ -79,6 +96,9 @@ export namespace FListResponse {
     type: 'file';
   }
 
+  /**
+   * File system directory representation
+   */
   export interface Dir {
     /**
      * Name of the directory
@@ -97,6 +117,9 @@ export namespace FListResponse {
   }
 }
 
+/**
+ * Response containing file content
+ */
 export interface FReadResponse {
   /**
    * Content of the file
@@ -104,6 +127,9 @@ export interface FReadResponse {
   content: string;
 }
 
+/**
+ * Response after writing file content
+ */
 export interface FWriteResponse {
   /**
    * Success message